src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.2.1
aiofiles==25.1.0
pydantic-settings==2.12.0
langchain>=0.2.0
langchain-community==0.4.1
sentence-transformers==5.1.2
openai==2.8.1
cohere==5.5.8
google-generativeai==0.8.5
qdrant-client>=1.10.1 
motor==3.7.1
pydantic-mongo>=3.1.0
SQLAlchemy==2.0.44
asyncpg==0.30.0
psycopg2-binary==2.9.11
alembic==1.17.2
psycopg>=3.1.18
pgvector==0.4.1
PyMuPDF>=1.26.6
nltk==3.9.1

# Monitoring and metrics
prometheus-client==0.21.1
starlette-exporter==0.23.0
fastapi-health==0.4.0